Adding one more layer
